% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/my_getdropbox.R
\name{my_setdropbox}
\alias{my_setdropbox}
\title{get dropbox folder}
\usage{
my_setdropbox(subfolder)
}
\arguments{
\item{subfolder}{Select a dropbox subfolder, can be eather "lists", "output" or "script"}
}
\description{
This function returns the location of my dropbox folder.
}
\examples{
my_setdropbox()
}
\keyword{dropbox}
\keyword{folder}

