% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/my_setwd.R
\name{my_setwd}
\alias{my_setwd}
\title{Set My Work Dir}
\usage{
my_setwd()
}
\arguments{
\item{this}{function takes no arguments}
}
\description{
This function sets my working directory.
}
\examples{
my_setwd()
}
\keyword{workdir}

