% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/name_computer.R
\name{name_computer}
\alias{name_computer}
\title{Computername}
\usage{
name_computer()
}
\arguments{
\item{this}{function takes no arguments}
}
\description{
Returns the name the computer.
}
\examples{
name_computer()
}
\keyword{computername}

